+++ examples/basic-request/main.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/coltiebaby/glaw"
+	"github.com/coltiebaby/glaw/league"
 )
 
 func main() {
@@ -19,8 +20,8 @@ func main() {
 		glaw.WithAPIToken(os.Getenv("LEAGUE_API_KEY")),
 	}
 
-	client, _ := glaw.NewClient(opts...)
-	req := glaw.ChampionRotationsRequest{
+	client, _ := league.NewClient(opts...)
+	req := league.ChampionRotationsRequest{
 		Region: glaw.REGION_NA,
 	}
