+++ league/client.go
@@ -1,0 +1,20 @@
+package league
+
+import (
+	"github.com/coltiebaby/glaw"
+)
+
+type Client struct {
+	*glaw.Client
+}
+
+func NewClient(opts ...glaw.Option) (*Client, error) {
+	client, err := glaw.NewClient(opts...)
+	if err != nil {
+		return nil, err
+	}
+
+	return &Client{
+		Client: client,
+	}, nil
+}
