+++ league/core/champion.go
@@ -1,0 +1,7 @@
+package core
+
+type ChampionInfo struct {
+	FreeChampionIds              []int `json:"freeChampionIds"`
+	FreeChampionIdsForNewPlayers []int `json:"freeChampionIdsForNewPlayers"`
+	MaxNewPlayerLevel            int   `json:"maxNewPlayerLevel"`
+}
