+++ league/core/masteries.go
@@ -1,0 +1,13 @@
+package core
+
+type ChampionMastery struct {
+	ChampionLevel                int    `json:"championLevel"`
+	ChestGranted                 bool   `json:"chestGranted"`
+	ChampionPoints               int    `json:"championPoints"`
+	ChampionPointsSinceLastLevel int    `json:"championPointsSinceLastLevel"`
+	ChampionPointsUntilNextLevel int    `json:"championPointsUntilNextLevel"`
+	SummonerID                   string `json:"summonerId"`
+	TokensEarned                 int    `json:"tokensEarned"`
+	ChampionID                   int    `json:"championId"`
+	LastPlayTime                 int64  `json:"lastPlayTime"`
+}
