+++ league/leagues.go
@@ -1,14 +1,17 @@
-package glaw
+package league
 
 import (
 	"context"
 	"fmt"
+
+	"github.com/coltiebaby/glaw"
+	"github.com/coltiebaby/glaw/league/core"
 )
 
 type QueueRequest struct {
 	Type   string
-	Queue  Queue
-	Region Region
+	Queue  core.Queue
+	Region glaw.Region
 }
 
 func (qr QueueRequest) String() string {
@@ -16,11 +19,11 @@ func (qr QueueRequest) String() string {
 	return fmt.Sprintf(template, qr.Type, qr.Queue)
 }
 
-func (c *Client) Queue(ctx context.Context, qr QueueRequest) (league League, err error) {
-	req := Request{
+func (c *Client) Queue(ctx context.Context, qr QueueRequest) (league core.League, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `league`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  qr.Region,
 		Uri:     qr.String(),
 	}
@@ -30,24 +33,24 @@ func (c *Client) Queue(ctx context.Context, qr QueueRequest) (league League, err
 		return league, err
 	}
 
-	err = ProcessResponse(resp, &league)
+	err = glaw.ProcessResponse(resp, &league)
 	return league, err
 }
 
 type LeagueRequest struct {
 	ID     string
-	Region Region
+	Region glaw.Region
 }
 
 func (lr LeagueRequest) String() string {
 	return fmt.Sprintf(`leagues/%s`, lr.ID)
 }
 
-func (c *Client) League(ctx context.Context, lr LeagueRequest) (league League, err error) {
-	req := Request{
+func (c *Client) League(ctx context.Context, lr LeagueRequest) (league core.League, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `league`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  lr.Region,
 		Uri:     lr.String(),
 	}
@@ -57,7 +60,7 @@ func (c *Client) League(ctx context.Context, lr LeagueRequest) (league League, e
 		return league, err
 	}
 
-	err = ProcessResponse(resp, &league)
+	err = glaw.ProcessResponse(resp, &league)
 	return league, err
 }
 
@@ -91,75 +94,3 @@ func (c *Client) League(ctx context.Context, lr LeagueRequest) (league League, e
 // 	entries, err = getEntries(c, fmt.Sprintf(`entries/by-summoner/%s`, id), -1)
 // 	return entries, err
 // }
-
-type Queue string
-
-const (
-	Challenger  = `challenger`
-	Master      = `master`
-	GrandMaster = `grandmaster`
-)
-
-const (
-	FLEX               Queue = `RANKED_FLEX`
-	SOLO               Queue = `RANKED_SOLO_5x5`
-	TEAM_FIGHT_TACTICS Queue = `RANKED_FLEX_TFT`
-	TWISTED_TREELINE   Queue = `RANKED_FLEX_TT`
-)
-
-type Tier string
-
-const (
-	// CHALLENGER Tier = "CHALLENGER"
-	DIAMOND  Tier = `DIAMOND`
-	PLATINUM Tier = `PLATINUM`
-	GOLD     Tier = `GOLD`
-	SILVER   Tier = `SILVER`
-	BRONZE   Tier = `BRONZE`
-	IRON     Tier = `IRON`
-)
-
-type Division string
-
-const (
-	ONE   Division = `I`
-	TWO   Division = `II`
-	THREE Division = `III`
-	FOUR  Division = `IV`
-)
-
-type Entry struct {
-	MiniSeries   MiniSeries
-	SummonerName string `json:"summonerName"`
-	HotStreak    bool   `json:"hotStreak"`
-	Wins         int    `json:"wins"`
-	Veteran      bool   `json:"veteran"`
-	Losses       int    `json:"losses"`
-	Rank         string `json:"rank"`
-	Inactive     bool   `json:"inactive"`
-	FreshBlood   bool   `json:"freshBlood"`
-	SummonerID   string `json:"summonerId"`
-	LeaguePoints int    `json:"leaguePoints"`
-}
-
-type MiniSeries struct {
-	Progress string `json:"progress"`
-	Losses   int    `json:"losses"`
-	Target   int    `json:"target"`
-	Wins     int    `json:"wins"`
-}
-
-type League struct {
-	Tier     string  `json:"tier"`
-	LeagueID string  `json:"leagueId"`
-	Entries  []Entry `json:"entries"`
-	Queue    string  `json:"queue"`
-	Name     string  `json:"name"`
-}
-
-// Has slightly more info than the Entry struct
-type LeagueEntry struct {
-	QueueType string `json:"queueType"`
-	Tier      string `json:"tier"`
-	Entry
-}
