+++ league/masteries.go
@@ -1,12 +1,15 @@
-package glaw
+package league
 
 import (
 	"context"
 	"fmt"
+
+	"github.com/coltiebaby/glaw"
+	"github.com/coltiebaby/glaw/league/core"
 )
 
 type ScoreRequest struct {
-	Region              Region
+	Region              glaw.Region
 	EncryptedSummonerID string
 }
 
@@ -15,10 +18,10 @@ func (sr ScoreRequest) String() string {
 }
 
 func (c *Client) ChampionScore(ctx context.Context, mr MasteryRequest) (score int, err error) {
-	req := Request{
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `champion-mastery`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  mr.Region,
 		Uri:     mr.String(),
 	}
@@ -28,12 +31,12 @@ func (c *Client) ChampionScore(ctx context.Context, mr MasteryRequest) (score in
 		return score, err
 	}
 
-	err = ProcessResponse(resp, &score)
+	err = glaw.ProcessResponse(resp, &score)
 	return score, err
 }
 
 type MasteryRequest struct {
-	Region              Region
+	Region              glaw.Region
 	EncryptedSummonerID string
 }
 
@@ -41,11 +44,11 @@ func (mr MasteryRequest) String() string {
 	return fmt.Sprintf("champion-masteries/by-summoner/%s", mr.EncryptedSummonerID)
 }
 
-func (c *Client) ChampionMasteries(ctx context.Context, mr MasteryRequest) (cm []ChampionMastery, err error) {
-	req := Request{
+func (c *Client) ChampionMasteries(ctx context.Context, mr MasteryRequest) (cm []core.ChampionMastery, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `champion-mastery`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  mr.Region,
 		Uri:     mr.String(),
 	}
@@ -55,12 +58,12 @@ func (c *Client) ChampionMasteries(ctx context.Context, mr MasteryRequest) (cm [
 		return cm, err
 	}
 
-	err = ProcessResponse(resp, &cm)
+	err = glaw.ProcessResponse(resp, &cm)
 	return cm, err
 }
 
 type MasteriesRequest struct {
-	Region              Region
+	Region              glaw.Region
 	EncryptedSummonerID string
 	ChampionID          int
 }
@@ -69,11 +72,11 @@ func (mr MasteriesRequest) String() string {
 	return fmt.Sprintf("champion-masteries/by-summoner/%s/by-champion/%d", mr.EncryptedSummonerID, mr.ChampionID)
 }
 
-func (c *Client) MasteriesByChampionId(ctx context.Context, mr MasteryRequest) (cm ChampionMastery, err error) {
-	req := Request{
+func (c *Client) MasteriesByChampionId(ctx context.Context, mr MasteryRequest) (cm core.ChampionMastery, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `champion-mastery`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  mr.Region,
 		Uri:     mr.String(),
 	}
@@ -83,18 +86,6 @@ func (c *Client) MasteriesByChampionId(ctx context.Context, mr MasteryRequest) (
 		return cm, err
 	}
 
-	err = ProcessResponse(resp, &cm)
+	err = glaw.ProcessResponse(resp, &cm)
 	return cm, err
-}
-
-type ChampionMastery struct {
-	ChampionLevel                int    `json:"championLevel"`
-	ChestGranted                 bool   `json:"chestGranted"`
-	ChampionPoints               int    `json:"championPoints"`
-	ChampionPointsSinceLastLevel int    `json:"championPointsSinceLastLevel"`
-	ChampionPointsUntilNextLevel int    `json:"championPointsUntilNextLevel"`
-	SummonerID                   string `json:"summonerId"`
-	TokensEarned                 int    `json:"tokensEarned"`
-	ChampionID                   int    `json:"championId"`
-	LastPlayTime                 int64  `json:"lastPlayTime"`
 }
