+++ league/match.go
@@ -1,24 +1,27 @@
-package glaw
+package league
 
 import (
 	"context"
 	"fmt"
+
+	"github.com/coltiebaby/glaw"
+	"github.com/coltiebaby/glaw/league/core"
 )
 
 type MatchRequest struct {
 	ID     string
-	Region Region
+	Region glaw.Region
 }
 
 func (mr MatchRequest) String() string {
 	return fmt.Sprintf("matches/%s", mr.ID)
 }
 
-func (c *Client) Match(ctx context.Context, mr MatchRequest) (matches MatchStorage, err error) {
-	req := Request{
+func (c *Client) Match(ctx context.Context, mr MatchRequest) (matches core.MatchStorage, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `match`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  mr.Region,
 		Uri:     mr.String(),
 	}
@@ -28,24 +31,24 @@ func (c *Client) Match(ctx context.Context, mr MatchRequest) (matches MatchStora
 		return matches, err
 	}
 
-	err = ProcessResponse(resp, &matches)
+	err = glaw.ProcessResponse(resp, &matches)
 	return matches, err
 }
 
 type MatchesRequest struct {
 	AccountID string
-	Region    Region
+	Region    glaw.Region
 }
 
 func (mr MatchesRequest) String() string {
 	return fmt.Sprintf("matchlists/by-account/%s", mr.AccountID)
 }
 
-func (c *Client) Matches(ctx context.Context, mr MatchRequest) (matches MatchStorage, err error) {
-	req := Request{
+func (c *Client) Matches(ctx context.Context, mr MatchRequest) (matches core.MatchStorage, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `match`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  mr.Region,
 		Uri:     mr.String(),
 	}
@@ -55,24 +58,24 @@ func (c *Client) Matches(ctx context.Context, mr MatchRequest) (matches MatchSto
 		return matches, err
 	}
 
-	err = ProcessResponse(resp, &matches)
+	err = glaw.ProcessResponse(resp, &matches)
 	return matches, err
 }
 
 type TimelineRequest struct {
 	ID     string
-	Region Region
+	Region glaw.Region
 }
 
 func (tr TimelineRequest) String() string {
 	return fmt.Sprintf("timelines/by-match/%s", tr.ID)
 }
 
-func (c *Client) Timeline(ctx context.Context, mr MatchRequest) (matches MatchStorage, err error) {
-	req := Request{
+func (c *Client) Timeline(ctx context.Context, mr MatchRequest) (matches core.MatchStorage, err error) {
+	req := glaw.Request{
 		Method:  `GET`,
 		Domain:  `match`,
-		Version: V4,
+		Version: glaw.V4,
 		Region:  mr.Region,
 		Uri:     mr.String(),
 	}
@@ -82,6 +85,6 @@ func (c *Client) Timeline(ctx context.Context, mr MatchRequest) (matches MatchSt
 		return matches, err
 	}
 
-	err = ProcessResponse(resp, &matches)
+	err = glaw.ProcessResponse(resp, &matches)
 	return matches, err
 }
